+++ filters/ConvolutionFilter.h
@@ -21,6 +21,7 @@
 
 #include "IFilter.h"
 #include "libHybridConv-0.1.1/libHybridConv_eapo.h"
+//#include "fftwrapper.h"
 
 #pragma AVRT_VTABLES_BEGIN
 class ConvolutionFilter : public IFilter
@@ -30,13 +31,14 @@ public:
 	virtual ~ConvolutionFilter();
 	bool getInPlace() override {return true;}
 	std::vector<std::wstring> initialize(float sampleRate, unsigned maxFrameCount, std::vector<std::wstring> channelNames) override;
-	void process(float** output, float** input, unsigned frameCount) override;
+	void process(double** output, double** input, unsigned frameCount) override;
 
 private:
 	void cleanup();
 
 	std::wstring filename;
 	HConvSingle* filters;
+	//FFTW_R2C_1D_Executor **filters;
 	unsigned channelCount;
 };
 #pragma AVRT_VTABLES_END
